chore: use same Algolia index name handling
