+++ packages/dnb-design-system-portal/src/uilib/search/searchQuery.js
@@ -5,7 +5,7 @@
 
 const getCurrentBranchName = require('current-git-branch')
 const { makeSlug } = require('../utils/slug')
-const { isCI } = require('ci-info')
+const { getIndexName, runQueriesWhen } = require('./searchHelpers')
 
 require('dotenv').config()
 
@@ -132,38 +132,6 @@ const flatten = (arr) =>
 const hasTitle = (r) => String(r.title || '').length > 0
 const hasSearch = (r) => String(r.search || '').length > 0
 const hasDescription = (r) => String(r.description || '').length > 0
-
-// Finds current index name for the Algolia search
-const getIndexName = (currentBranch) => {
-  if (process.env.NODE_ENV !== 'production' || !isCI) {
-    return 'dev_eufemia_docs'
-  }
-
-  if (currentBranch !== 'release') {
-    return 'beta_eufemia_docs'
-  }
-
-  return 'prod_eufemia_docs'
-}
-
-const runQueriesWhen = (currentBranch) => {
-  if ((process.env.ALGOLIA_API_KEY || '').length === 0) {
-    console.info(
-      'If you want to submit searchable data to Algolia, you need to request access keys and put them in a local .env file.'
-    )
-    return false
-  }
-
-  if (isCI) {
-    return /^(release|beta|portal)$/.test(currentBranch)
-  }
-
-  if (process.env.NODE_ENV === 'production') {
-    return true
-  }
-
-  return false
-}
 
 const currentBranch = getCurrentBranchName()
